// AngularCinema/src/models/IFilm.ts
import { IRoom } from "./IRoom";

export interface IFilm {
  id: number
  title: string
  screeningTime: number
  isShowing: boolean
        // public int ID { get; set; }
        // public string Title { get; set; }
        // // In minutes.
        // public int ScreeningTime { get; set; }

        // public bool IsShowing { get; set; }
}